tests/requirements.txt
# Onnx conversion
tf2onnx~=1.9.1
onnxruntime~=1.8.1

# Tensorflow Model Optimization Toolkit
tensorflow-model-optimization~=0.6.0

# System info
psutil~=5.8.0
